SQL実行機能 * 実装見直し。 * common最新化

+++ src/main/java/cherry/sqlapp/controller/sqltool/clause/SqltoolClauseForm.java
@@ -64,8 +64,8 @@ public class SqltoolClauseForm implements Serializable {
 
 	private Integer lockVersion;
 
-	private int pageNo;
+	private long pageNo;
 
-	private int pageSz;
+	private long pageSz;
 
 }
